ski-api/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
supabase==2.4.0
tflite-runtime==2.14.0
numpy==1.26.4
